refactored refinesnsolver.h
